Source reconstruction for the 64/256K @ TECO.
The @ TECO binary is from the . directory on many backup tapes going
all the way back to 1971.  The LPT 11 and MACTAP F68 sources come
unchanged from Peter Samson's "RandomSystem" DECtape.  The TECO DUMMY
file is close to the undated SYSENG; TECODM 2 file.

The source files can be assembled with MIDAS 73, and linked with the
Muddle version of STINK.  The link order is TECO DUMMY, LPT, MACTAP,
TECO.

This results in a binary that is identical to the original @ TECO with
a few exceptions:

- @ TECO has a loader in 20-37, and something in 776000-777777.
- MOBY is set to 174000 for 64K core, which results in MACDMP being
  1777400 and MACCR being 177777.  The @ TECO binary has it set to
  DSKDMP at 777700.
- The buffer size in MEMSIZ is not the same.
- CONSTANTS has one additional zero at the end, shifting the rest
  of the memory image up by one word.